backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
pdfplumber==0.11.4
google-generativeai==0.8.3
httpx==0.27.2
